Added a fake API interface
We'll use this to make it look like we're making API calls, using proper promises and response payloads.
bootstrap.js has most of the store.js code but cleaned-up and nicer to use and is intended to replace it.
bootstrap.js:generateSocialNetwork() is will create a socially connected group of people which is consistent in terms of connections between people (unlike store.js which is completely random). This allows more realistic testing
api.js leverages the generated social network to simulate an API