Added store/users to vuex
All user information is stored in this store including extra meta data such as who the current user is etc...
There can only be at most one 'activeUser' which is generally the user for the currently viewed profile page.
Since 'me' or '<userId>' are both valid, we also store this information in the store. The meUserId is vital and should be loaded into the store before anything else (basically, requires a fetch of the user's own profile)